requirements.txt
python-telegram-bot[job-queue]==20.7
libtorrent==2.0.11
pydrive2==1.21.3